requirements.txt
python-decouple
speechrecognition
pyttsx3
openai
disutils
pipwin
pyaudio